+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 
 ## [Unreleased]
 
+### Added
+
+- ACL proxy application
+
 ## [0.3.0] - 2022-01-27
 
 ### Added